Refactor Default Routes
app_server
  controllers
  models
  routes
  views

close https://github.com/knightoftheoldcode/rehome/issues/11
