+++ app_server/controllers/main.js
@@ -1,0 +1,8 @@
+/* GET home page. */
+const index = function(req, res, next) {
+  res.render('index', { title: 'rehome' });
+};
+
+module.exports = {
+  index
+}
