+++ app_server/routes/index.js
@@ -1,9 +1,8 @@
 var express = require('express');
 var router = express.Router();
+const ctrlMain = require('../controllers/main')
 
 /* GET home page. */
-router.get('/', function(req, res, next) {
-  res.render('index', { title: 'rehome' });
-});
+router.get('/', ctrlMain.index);
 
 module.exports = router;
